refactor : supabase api  mangement and change file name
