+++ src/app/(root)/detail/[id]/page.tsx
@@ -1,6 +1,6 @@
-import { fetchBakeryDetails } from "@/app/api/supabase/(detail)/route";
+import { fetchBakeryDetails } from "@/app/api/supabase/detail/route";
 import CommentList from "@/components/comment/CommentList";
-import SkeletonMap from "@/components/commons/Skeleton/SkeletonMap";
+import SkeletonMap from "@/components/commons/skeleton/SkeletonMap";
 import StoreInformation from "@/components/detailMap/StoreInformation";
 import { Metadata } from "next";
 import dynamic from "next/dynamic";
