+++ src/app/(root)/modify-profile/page.tsx
@@ -2,7 +2,7 @@
 
 // 병렬 라우팅을 위한 페이지입니다
 const Page = () => {
-    return <>this is modifyprofile page</>;
-}
+  return <>this is modifyprofile page</>;
+};
 
 export default Page;
