+++ src/app/(root)/page.tsx
@@ -1,4 +1,4 @@
-import { BakeryList } from "@/components/BakeryList/BakeryList";
+import { BakeryList } from "@/components/bakeries/BakeryList";
 import SearchBar from "@/components/commons/SearchBar";
 import UpButton from "@/components/commons/UpButton";
 import BouncingBread from "@/components/main/BouncingBread";
