+++ src/app/(root)/signup/page.tsx
@@ -5,10 +5,10 @@ import Link from "next/link";
 import { useRouter } from "next/navigation";
 import { useRef, useState } from "react";
 import { v4 as uuidv4 } from "uuid";
-import LogoBread from "../../../../public/image/breads/LogoBread.png";
 import { uploadImage } from "../../../supabase/utils/makeimageUrl";
 import { signUp, updateUserProfile } from "../../api/supabase/auth/route";
 import nookies from "nookies";
+import LogoBread from "../../../../public/image/breads/LogoBread.png";
 
 const SignupPage = () => {
   const [email, setEmail] = useState("");
@@ -19,7 +19,6 @@ const SignupPage = () => {
   const [profileUrl, setProfileUrl] = useState<string | null>(null);
   const router = useRouter();
   const newUuid = uuidv4();
-
   const fileInputRef = useRef<HTMLInputElement>(null);
 
   // 회원가입 처리 함수
