+++ src/app/api/supabase/detail/route.ts
