+++ src/app/api/supabase/like/route.ts
@@ -47,3 +47,19 @@ export const toggleLikeStatus = async (isLiked: boolean, userId: string, bakeryI
     return false;
   }
 };
+// `userId`를 사용하여 좋아요한 bakery의 ID 리스트를 가져오기
+export const getUserLikedBakeryIds = async (userId: string) => {
+  try {
+    const { data, error } = await supabase.from("like").select("bakery_id").eq("user_id", userId);
+
+    if (error) {
+      console.error("Failed to fetch liked bakeries:", error);
+      throw error;
+    }
+
+    return data?.map((row: any) => row.bakery_id) || [];
+  } catch (error) {
+    console.error("Error in getUserLikedBakeryIds:", error);
+    throw error;
+  }
+};
