+++ src/components/bakeries/BakeryList.tsx
@@ -1,35 +1,26 @@
 "use client";
 
-import { createClient } from "@/supabase/client";
-import { Bakery } from "@/types/bakery";
 import Link from "next/link";
 import { useEffect, useState } from "react";
+import { createClient } from "@/supabase/client";
+import { Bakery } from "@/types/bakery";
 import { BakeryCard } from "../commons/BakeryCard";
+import { bakeryItem } from "@/app/api/supabase/bakery/route";
 
-interface BakeryListProp {
+type BakeryListProp = {
   searchedBakeries?: Bakery[];
-}
+};
 
 export const BakeryList = ({ searchedBakeries }: BakeryListProp) => {
   const [breads, setBreads] = useState<Bakery[]>([]);
-  const [error, setError] = useState<string | null>(null);
 
   useEffect(() => {
     if (searchedBakeries) {
       setBreads(searchedBakeries);
     } else {
-      const supabase = createClient();
       const fetchBreads = async () => {
-        try {
-          const { data, error } = await supabase.from("bakery").select("*").order("sort_id", { ascending: true });
-
-          if (error) {
-            throw new Error(error.message);
-          }
-          setBreads((data as Bakery[]) || []);
-        } catch (error) {
-          console.error("목록을 불러오는 중 오류가 발생했습니다.", error);
-        }
+        const data = await bakeryItem();
+        setBreads((data as Bakery[]) || []);
       };
       fetchBreads();
     }
