+++ src/components/detailMap/KakaoMap.tsx
@@ -6,7 +6,7 @@ import Image from "next/image";
 import Link from "next/link";
 import { KakaoMapProps, GeocoderResult, WindowWithKakao } from "@/types/map";
 import { loadKakaoMapScript } from "@/app/api/kakao/route";
-import SkeletonMap from "@/components/commons/Skeleton/SkeletonMap";
+import SkeletonMap from "@/components/commons/skeleton/SkeletonMap";
 
 const KakaoMap: React.FC<KakaoMapProps> = ({ name, address }) => {
   const [isLoaded, setIsLoaded] = useState(false);
